break up sourcecode for spatial package into multiple files while i was in there i: * tweaked the node test helper to ensure that errors arent swallowed * limited travis to running node tests till i can figure out what is wrong with browser env * bumped babel deps to resolve a nag
was it just that?

lint for the lint gods

one more
