+++ package.json
@@ -37,9 +37,9 @@
   "homepage": "https://github.com/terraformer-js/terraformer#readme",
   "devDependencies": {
     "@babel/cli": "^7.2.3",
-    "@babel/core": "^7.2.2",
+    "@babel/core": "^7.9.0",
     "@babel/node": "^7.2.2",
-    "@babel/preset-env": "^7.2.3",
+    "@babel/preset-env": "^7.9.5",
     "babelify": "^10.0.0",
     "browserify": "^16.0.0",
     "faucet": "0.0.1",
