+++ packages/spatial/contains.js
@@ -1,0 +1,3 @@
+import { within } from './within';
+
+export const contains = (geoJSON, comparisonGeoJSON) => within(comparisonGeoJSON, geoJSON);
