+++ test-helper.js
@@ -20,9 +20,8 @@ for (var i = 0; i < packages.length; i++) {
   for (var j = 0; j < tests.length; j++) {
     // use babel to transpile the source and pass each test to the Node.js CLI
     exec(`babel-node ${packages[i]}/test/${tests[j]} [ babelify --presets @babel/preset-env ] | faucet`, function (err, res) {
-      if (!err) {
-        console.log(res);
-      }
+      if (err) console.error(err);
+      if (res) console.log(res);
     });
   }
 }
